+++ src/main/java/org/vafer/jdeb/maven/DebMojo.java
@@ -93,6 +93,13 @@ public class DebMojo extends AbstractPluginMojo {
      */
     private String compression;
 
+
+	/** Boolean option to attach the artifact to the project 
+	 *  
+	 *  @parameter default-value="true"
+	 */
+	private String attach;
+	
     /**
      * The location where all package files will be installed. By default, all
      * packages are installed in /opt (see the FHS here:
@@ -266,8 +273,12 @@ public class DebMojo extends AbstractPluginMojo {
                 debMaker.setCompression(compression);
                 debMaker.makeDeb();
 
-                getLog().info("Attaching created debian archive " + debFile);
-                projectHelper.attachArtifact(getProject(), type, classifier, debFile);
+                // Always attach unless explicitly set to false
+                if (attach == null || attach.compareToIgnoreCase("false") != 0)
+                {
+                    getLog().info("Attaching created debian archive " + debFile);
+                    projectHelper.attachArtifact(getProject(), type, classifier, debFile);
+                }
 
             } catch (PackagingException e) {
                 getLog().error("Failed to create debian package " + debFile, e);